refactor: move remaining McpContext getters
