+++ src/tools/webmcp.ts
@@ -19,7 +19,7 @@ export const listWebMcpTools = definePageTool({
   schema: {},
   blockedByDialog: false,
   verifyFilesSchema: [],
-  handler: async (_request, response, _context) => {
+  handler: async (_request, response) => {
     response.setListWebMcpTools();
   },
 });
